Chore:Layoutcontext fixes for post
