+++ src/app/publicaciones/layout.tsx
@@ -1,32 +1,8 @@
 "use client";
 
-import { createContext, useContext, useState } from "react";
+import { useMyContext } from "~/utils/posts/layoutContext";
 
-interface MyContextType {
-  value: boolean;
-  setValue: React.Dispatch<React.SetStateAction<boolean>>;
-}
 
-const MyContext = createContext<MyContextType | undefined>(undefined);
-
-export const MyProvider = ({ children }: { children: React.ReactNode }) => {
-  const [value, setValue] = useState<boolean>(true);
-
-  return (
-    <MyContext.Provider value={{ value, setValue }}>
-      {children}
-    </MyContext.Provider>
-  );
-};
-
-// Custom hook for easy access to the context
-export const useMyContext = (): MyContextType => {
-  const context = useContext(MyContext);
-  if (!context) {
-    throw new Error("useMyContext must be used within a MyProvider");
-  }
-  return context;
-};
 
 export default function PublicacionesLayout({
   children,
